+++ Controls/RoundTextBox.cs
@@ -1,0 +1,30 @@
+using System;
+using System.Drawing;
+using System.Windows.Forms;
+
+namespace filmhub.Controls
+{
+
+    public class RoundTextBox : TextBox
+    {
+        [System.Runtime.InteropServices.DllImport("gdi32.dll", EntryPoint = "CreateRoundRectRgn")]
+        private static extern IntPtr CreateRoundRectRgn
+        (
+            int nLeftRect, // X-coordinate of upper-left corner or padding at start
+            int nTopRect, // Y-coordinate of upper-left corner or padding at the top of the textbox
+            int nRightRect, // X-coordinate of lower-right corner or Width of the object
+            int nBottomRect, // Y-coordinate of lower-right corner or Height of the object
+            //RADIUS, how round do you want it to be?
+            int nheightRect, //height of ellipse 
+            int nweightRect //width of ellipse
+        );
+
+        protected override void OnCreateControl()
+        {
+            base.OnCreateControl();
+            Region =
+                Region.FromHrgn(CreateRoundRectRgn(2, 3, Width, Height, 
+                    15,15));
+        }
+    }
+}
